Move getUrl functionality to separate helper file
To keep the component a bit more tidy
