+++ src/components/cover/cover.tsx
@@ -1,18 +1,12 @@
 import React, { useCallback, useState } from "react";
 import clsx from "clsx";
-import { first } from "lodash";
 import { useGetCoverCollection } from "../../core/cover-service-api/cover-service";
-import {
-  Cover as CoverType,
-  CoverImageUrls,
-  GetCoverCollectionType
-} from "../../core/cover-service-api/model";
+import { GetCoverCollectionType } from "../../core/cover-service-api/model";
 import { Pid } from "../../core/utils/types/ids";
 import LinkNoStyle from "../atoms/links/LinkNoStyle";
 import CoverImage from "./cover-image";
 import { Manifestation } from "../../core/utils/types/entities";
-
-type CoverServiceSizes = keyof CoverImageUrls;
+import { getCoverUrl } from "./helper";
 
 export type CoverProps = {
   animate: boolean;
@@ -25,46 +19,6 @@ export type CoverProps = {
   idType?: GetCoverCollectionType;
   shadow?: "small" | "medium";
   linkAriaLabelledBy?: string;
-};
-
-const getUrl = (cover: CoverType, size: CoverServiceSizes) =>
-  cover.imageUrls?.[size]?.url;
-
-const getCoverUrl = ({
-  coverData,
-  bestRepresentation,
-  size
-}: {
-  coverData: CoverType[] | null | undefined;
-  bestRepresentation: CoverProps["bestRepresentation"];
-  size: CoverServiceSizes;
-}) => {
-  if (!coverData) {
-    return null;
-  }
-
-  const firstCover = first(coverData);
-
-  if (!bestRepresentation && firstCover && getUrl(firstCover, size)) {
-    return getUrl(firstCover, size);
-  }
-
-  const bestRepresentationCover = first(
-    coverData.filter(
-      (item: CoverType) =>
-        bestRepresentation && item.id === bestRepresentation.pid
-    )
-  );
-
-  if (bestRepresentationCover && getUrl(bestRepresentationCover, size)) {
-    return getUrl(bestRepresentationCover, size);
-  }
-
-  if (firstCover && getUrl(firstCover, size)) {
-    return getUrl(firstCover, size);
-  }
-
-  return null;
 };
 
 export const Cover = ({
